web-admin/requirements.txt
# Web Admin Interface Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-BasicAuth>=0.2.0
Flask-Caching>=2.0.2
Flask-Login>=0.6.2
PyYAML>=6.0
Jinja2>=3.1.2
requests>=2.31.0
gunicorn>=21.2.0
docker>=6.1.3
APScheduler>=3.10.1
tqdm>=4.65.0
beautifulsoup4>=4.12.0